adds logic to add multiple code/preview windows, creates CodeCell wrapper component for Editor & Preview
