+++ src/components/code-cell/code-cell.tsx
@@ -1,0 +1,28 @@
+import { useState } from 'react'
+import { CodeEditor, Preview } from '..'
+import { bundle } from '../../bundler'
+
+const CodeCell = () => {
+  const [input, setInput] = useState('')
+  const [code, setCode] = useState('')
+
+  const onClick = async () => {
+    const output = await bundle(input)
+    setCode(output)
+  }
+
+  return (
+    <div>
+      <CodeEditor
+        initialValue='// Start writing some code!'
+        onChange={val => setInput(val)}
+      />
+      <div>
+        <button onClick={onClick}>Submit</button>
+      </div>
+      <Preview code={code} />
+    </div>
+  )
+}
+
+export default CodeCell
