+++ src/components/index.ts
@@ -1,2 +1,3 @@
 export { default as CodeEditor } from './code-editor/code-editor'
 export { default as Preview } from './preview/preview'
+export { default as CodeCell } from './code-cell/code-cell'
